backend/requirements.txt
fastapi==0.124.2
uvicorn==0.39.0
httpx==0.28.1
python-dotenv==1.2.1
python-multipart==0.0.20
requests==2.32.5
elasticsearch==7.13.4